cast dot to int in abs() call, required for STAPL because STLport does not support abs(double)
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/pmpl/trunk@850 5e2c2758-810a-0410-8eb3-8e4e92d4004e

+++ OBPRM/Cfg_2D.cpp
@@ -258,7 +258,7 @@ Cfg_2D::GetCfgByOverlappingNormal(Environment* env,CollisionDetection* cd,
   // Overlap these two normals, solve for alpha, beta, gamma of FixedXYZ rotation.
   Orientation orient;
   double dot = robotNormal.dotProduct(obstNormal);
-  if(abs(dot) == 1) { // two normals parallel to each other.
+  if(abs((int)dot) == 1) { // two normals parallel to each other.
     orient = Orientation(IdentityMatrix);
   } else {
     double cV = sqrt((1+dot)/2.0);
